Refactoring.
Refactor classnames to make full use of namespaces for smoother paths.
Remove the Action Scheduler dependency, which provides no added benefit over a regular WP_Cron implementation at this time.
Refactor the sync hooks to be REST endpoints with authentication needs, passed over `apiFetch` for consistency with the rest of the codebase.