Pre-load Imageshop with permalinks.
This sends a bulk request for any queued permalinks at the end of a request to have them built.

A more performant approach would be to offload this to WP_Cron, but since we can not guarantee that WP_Cron is available, working like it should, and able to be invoked in a timely enough manner, we instead settle for a shutdown handler. This means that load times are 1-3 seconds, instead of 200ms, but this is an acceptable tradeoff in favor of guaranteed user experiences.