I18n: Improve language detection and manual overrides.
This improves the detection of the default language of a site to use as the default lookup language when making calls to the Imageshop API, ensuring that the site language is used as the basis for fields such as: description, alt texts, and captions.

A new filter selection for languages has also been introduced, allowing the editor to switch which language is being returned from Imageshop calls, regardless of their sites language options.

And finally, media modal languages have had their strings passed through translation functions, allowing for future translations to improve the overall experience.