Add fallback handling for media missing attachment IDs.
It is possible when an attachment is called outside the primary loop to be passed with an attachment ID of `0`, causing the srcset generator to not have any reference points.

This change implements a check to validate an attachment is passed, and if not, tries to extrapolate it based on the original URL in the markup.