Images: Add fallback solution for determining primary image dimensions
This introduces a fallback means to identify primary images for extracting image dimensions used to calculate aspect ratios.

Along with it, a new means of individually flushing caches and regenerating permalinks for individual images is introduced for those rare cases where a missing primary image has led to an undersized aspect ratio.