Attachments: Unify lookup function for original files.
The original file is part of the Attachment objects list of sub documents, but it may be declared in a variety of ways. By unifying the identification of this file, we can more easily add rules to catch edge cases, and ensure that all instances use the same media file.