Attachments: Account for cross-referenced content in WPML sites.
There are scenarios where content, such as `OG` attributes in a WPML site are used across multiple languages. This led to multiple fallback mechanics kicking in as the URL would not match any valid address for a given language. This allows for a non-URL string match lookup as the initial fallback to reduce this load, and increase the chance of finding a media match.