Attachments: Add dynamic handling of media captions.
If a media caption is dynamically queried, instead of relying on the data provided by the media dialog when inserted into content, it would in theory be empty, as no local copy exists.

This introduces captions as transient values that can be generated based on the description, and credits, from the Imageshop manager. The caption is stored in a transient for up to 1 week, but may be manually cleared per image as needed by flushing the Imageshop cache.