Force an image mime-type, even if no valid one exists.
Imageshop creates JPEG variants of image formats that WordPress by default does not support. This change provides a fallback value where Imageshop has declared the media item an image, and created the appropriate image variation of it, but may have an original mime-type that WordPress does not support, causing issues with the attachment post type.