Restrict the fallback metadata lookups.
The fallback method to find images if there's no attachment ID is to query the postmeta table for image URLs that match a given pattern. As the postmeta table does not have an index for the values (as it shouldn't), we can slightly speed up this process by at least only looking into items where the corresponding meta key is that of attachment metadata.